Migrate to ok-ex-button
