+++ packages/angular-workspace/example-client-app/src/app/customapp/customapp.component.html
@@ -37,7 +37,7 @@
         <example-chat-conversation-section></example-chat-conversation-section>
         <example-icons-section></example-icons-section>
         <example-rectangle-section></example-rectangle-section>
-        <example-ok-button-section></example-ok-button-section>
+        <example-ex-section></example-ex-section>
         <example-fv-section></example-fv-section>
     </div>
 </div>
