+++ packages/angular-workspace/example-client-app/src/app/customapp/ex/ex-button-section.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'example-ex-button-section',
+    template: `
+        <example-sub-container label="Button (Ok)">
+            <ok-ex-button>Ok</ok-ex-button>
+        </example-sub-container>
+    `,
+    standalone: false
+})
+export class ExButtonSectionComponent {}
