+++ packages/angular-workspace/example-client-app/src/app/customapp/ex/ex-section.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'example-ex-section',
+    template: `
+        <example-ex-button-section></example-ex-button-section>
+    `,
+    standalone: false
+})
+export class ExSectionComponent {}
