--- packages/angular-workspace/example-client-app/src/app/customapp/ok-button-section.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'example-ok-button-section',
-    template: `
-        <example-sub-container label="Button (Ok)">
-            <ok-button>Ok</ok-button>
-        </example-sub-container>
-    `,
-    standalone: false
-})
-export class OkButtonSectionComponent {}
