--- packages/angular-workspace/ok-angular/button/ng-package.json
@@ -1,6 +1,0 @@
-{
-    "$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
-    "lib": {
-      "entryFile": "public-api.ts"
-    }
-}
