--- packages/angular-workspace/ok-angular/button/public-api.ts
@@ -1,2 +1,0 @@
-export * from './ok-button.directive';
-export * from './ok-button.module';
