+++ packages/angular-workspace/ok-angular/ex/button/ok-ex-button.directive.ts
@@ -1,0 +1,14 @@
+import { Directive } from '@angular/core';
+import { type ExButton, exButtonTag } from '@ni/ok-components/dist/esm/ex/button';
+
+export type { ExButton };
+export { exButtonTag };
+
+/**
+ * Directive to provide Angular integration for the button.
+ */
+@Directive({
+    selector: 'ok-ex-button',
+    standalone: false
+})
+export class OkExButtonDirective { }
