+++ packages/angular-workspace/ok-angular/ex/button/tests/ok-ex-button.directive.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+import { OkExButtonModule } from '../ok-ex-button.module';
+
+describe('Ok ex button', () => {
+    describe('module', () => {
+        beforeEach(() => {
+            TestBed.configureTestingModule({
+                imports: [OkExButtonModule]
+            });
+        });
+
+        it('custom element is defined', () => {
+            expect(customElements.get('ok-ex-button')).not.toBeUndefined();
+        });
+    });
+});
