--- packages/angular-workspace/ok-angular/fv/accordion-item/fv-accordion-item.module.ts
@@ -1,12 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { OkFvAccordionItemDirective } from './fv-accordion-item.directive';
-
-import '@ni/ok-components/dist/esm/fv/accordion-item';
-
-@NgModule({
-    declarations: [OkFvAccordionItemDirective],
-    imports: [CommonModule],
-    exports: [OkFvAccordionItemDirective]
-})
-export class OkFvAccordionItemModule { }
