+++ packages/angular-workspace/ok-angular/fv/accordion-item/ok-fv-accordion-item.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { OkFvAccordionItemDirective } from './ok-fv-accordion-item.directive';
+
+import '@ni/ok-components/dist/esm/fv/accordion-item';
+
+@NgModule({
+    declarations: [OkFvAccordionItemDirective],
+    imports: [CommonModule],
+    exports: [OkFvAccordionItemDirective]
+})
+export class OkFvAccordionItemModule { }
