+++ packages/blazor-workspace/Examples/Demo.Shared/Pages/ComponentsDemo.razor
@@ -43,7 +43,7 @@
         <WaferMapSection />
         <ChatConversationSection />
         <RectangleSection />
-        <OkButtonSection />
+        <ExSection />
         <FvSection />
     </div>
 </div>
