+++ packages/blazor-workspace/Examples/Demo.Shared/Pages/Sections/Ex/ExButtonSection.razor
@@ -1,0 +1,7 @@
+@namespace Demo.Shared.Pages.Sections
+
+<div>
+    <SubContainer Label="Button (Ok)">
+        <OkExButton>Ok</OkExButton>
+    </SubContainer>
+</div>
