+++ packages/blazor-workspace/Examples/Demo.Shared/Pages/Sections/Ex/ExSection.razor
@@ -1,0 +1,3 @@
+@namespace Demo.Shared.Pages.Sections
+
+<ExButtonSection />
