--- packages/blazor-workspace/Examples/Demo.Shared/Pages/Sections/OkButtonSection.razor
@@ -1,7 +1,0 @@
-@namespace Demo.Shared.Pages.Sections
-
-<div>
-    <SubContainer Label="Button (Ok)">
-        <OkButton>Ok</OkButton>
-    </SubContainer>
-</div>
