--- packages/blazor-workspace/OkBlazor/Source/Button/OkButton.razor
@@ -1,5 +1,0 @@
-@namespace OkBlazor
-<ok-button
-    @attributes="AdditionalAttributes">
-    @ChildContent
-</ok-button>
