--- packages/blazor-workspace/OkBlazor/Source/Button/OkButton.razor.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.AspNetCore.Components;
-
-namespace OkBlazor;
-
-public partial class OkButton : ComponentBase
-{
-    /// <summary>
-    /// The child content of the element.
-    /// </summary>
-    [Parameter]
-    public RenderFragment? ChildContent { get; set; }
-
-    /// <summary>
-    /// Any additional attributes that did not match known properties.
-    /// </summary>
-    [Parameter(CaptureUnmatchedValues = true)]
-    public IDictionary<string, object>? AdditionalAttributes { get; set; }
-}
