+++ packages/blazor-workspace/OkBlazor/Source/Ex/Button/OkExButton.razor
@@ -1,0 +1,5 @@
+@namespace OkBlazor
+<ok-ex-button
+    @attributes="AdditionalAttributes">
+    @ChildContent
+</ok-ex-button>
