+++ packages/ok-components/src/ex/all-ex.ts
@@ -1,0 +1,1 @@
+import './button';
