--- packages/react-workspace/ok-react/src/button/index.ts
@@ -1,4 +1,0 @@
-import { Button } from '@ni/ok-components/dist/esm/button';
-import { wrap } from '../utilities/react-wrapper';
-
-export const OkButton = wrap(Button);
