+++ packages/react-workspace/ok-react/src/ex/button/index.ts
@@ -1,0 +1,4 @@
+import { ExButton } from '@ni/ok-components/dist/esm/ex/button';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const OkExButton = wrap(ExButton);
