+++ packages/react-workspace/react-client-app/src/components/Ex/ExButtonSection.tsx
@@ -1,0 +1,10 @@
+import { OkExButton } from '@ni/ok-react/ex/button';
+import { SubContainer } from '../SubContainer';
+
+export function ExButtonSection(): React.JSX.Element {
+    return (
+        <SubContainer label="Button (Ok)">
+            <OkExButton>Ok</OkExButton>
+        </SubContainer>
+    );
+}
