+++ packages/react-workspace/react-client-app/src/components/Ex/ExSection.tsx
@@ -1,0 +1,7 @@
+import { ExButtonSection } from './ExButtonSection';
+
+export function ExSection(): React.JSX.Element {
+    return (
+        <ExButtonSection></ExButtonSection>
+    );
+}
