--- packages/storybook/src/ok/button/button.mdx
@@ -1,19 +1,0 @@
-import { Controls, Canvas, Meta, Title } from '@storybook/addon-docs/blocks';
-import * as buttonStories from './button.stories';
-import ComponentApisLink from '../../docs/component-apis-link.mdx';
-
-<Meta of={buttonStories} />
-<Title of={buttonStories} />
-
-An Button is a temporary, demo Button component.
-
-<Canvas of={buttonStories.button} />
-
-## API
-
-<Controls of={buttonStories.button} />
-<ComponentApisLink />
-
-## Usage
-
-Don't.
