+++ packages/storybook/src/ok/ex/button/ex-button.mdx
@@ -1,0 +1,19 @@
+import { Controls, Canvas, Meta, Title } from '@storybook/addon-docs/blocks';
+import * as exButtonStories from './ex-button.stories';
+import ComponentApisLink from '../../../docs/component-apis-link.mdx';
+
+<Meta of={exButtonStories} />
+<Title of={exButtonStories} />
+
+An Ex Button is a temporary, demo Button component.
+
+<Canvas of={exButtonStories.exButton} />
+
+## API
+
+<Controls of={exButtonStories.exButton} />
+<ComponentApisLink />
+
+## Usage
+
+Don't.
